kjoy/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
cron-descriptor==1.4.0
cryptography==41.0.3
decorator==5.1.1
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.3
django-redis==5.3.0
django-rest-knox==4.2.0
django-timezone-field==6.0
djangorestframework==3.14.0
fake-useragent==1.2.1
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
Markdown==3.4.4
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
